Cirrus: Implement new retrospective task
For reasons such as releasing software or notifying people
about various conditions, a method is required to detect
and execute commands/scripts when all Cirrus-CI operations
have definitevly concluded.  However, due to the nature of
automation, there are occasional and unpredictable needs
to manually perform these activities.

Given existing familiarity and dependence on Cirrus-CI, add a
'retrospective' task with a 'manual' trigger and no dependencies.
Remove other notification/monitoring operations in anticipation
this task will take them over in a future commit.  Add a
github action that will atomicly and trigger the new task
on branch and tag push, under one of two conditions:

* The new github Action detects the Cirrus-CI 'success' task
  finished running and was not manually canceled.
* A person pressed the manual 'trigger' button on the new
  'retrospective' task, from the status page, on any build.

Also: E-mail me immediatly if there is ever a problem with the
      magic post-cirrus triggering automation.

Signed-off-by: Chris Evich <cevich@redhat.com>