podman: split env variables in env and overrides
There are three different priorities for applying env variables:

1) environment/config file environment variables
2) image's config
3) user overrides (--env)

The first and third kind are known to the client, while the image's
config is handled by the backend.

We need to carry the variables having different priorities in
different maps, so the backend can correctly build the final
environment configuration.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>